[tests] Extract MockSnapshot
For reuse in other tests.
